Update LMS
added a new function: Example books, this is simply addbooks but without the message to the user. Because this is used by the program to add books when it starts. So that the system can display books at all times.

Also added a new for loop during the add book in case the book being added is already in the system. Skipping over it with a short message and continuing to the user menu.